+++ crates/aztibase-network/src/behaviour.rs
@@ -2,7 +2,7 @@ use libp2p::connection_limits;
 use libp2p::kad::store::MemoryStore;
 use libp2p::request_response;
 use libp2p::swarm::NetworkBehaviour;
-use libp2p::{gossipsub, kad, mdns};
+use libp2p::{autonat, gossipsub, kad, mdns, relay};
 
 use crate::light_sync::LightSyncCodec;
 
@@ -13,4 +13,6 @@ pub struct AztibaseBehaviour {
     pub mdns: mdns::tokio::Behaviour,
     pub connection_limits: connection_limits::Behaviour,
     pub light_sync: request_response::Behaviour<LightSyncCodec>,
+    pub autonat: autonat::Behaviour,
+    pub relay_client: relay::client::Behaviour,
 }